Refactored speedtest script
Also while I added support for speedof.me they seem to be looking at the
phantomjs headers and fail any request I send. I left the code in but it
is unusable.

You can now override the script to go to a specific site by adding the
site to the arugments of the script.