Removed verizon support
So after some work it seems to be that the test failing a lot is not a
me issue but seems to be an issue between phantomjs and them. Since
their website gets the speed you pay for and could easily manipulate
your results I have no issue removing them.